Display Wind Outlook for Specified Location
- Created LandingVC which takes in location to fetch forecast
- WindOutlookVC to handle displaying daily wind forecast
- WeatherServicePublisher to handle fetching wind forecast
- DayOutlookModel & DayOutlookTableViewCell created & displaying
- Implemented Observer Design Pattern for notifying weather updates

User is presented with landing screen where they can input the location
for which they want to receive a daily wind forecast for. Upon clicking,
WeatherServicePublisher will fetch wind forecast for input location and
notify WindOutlookViewController when it receives a WebServiceResult via
Observer pattern. As of now, WebServiceResult is not handled, instead,
DayOutlookTableViewCells are displayed with stubbed data.

To-Do:
- Handle WebServiceResult from fetchAllWeatherForecast(lat:lon:) call
- Use returned data to populate each DayOutlookTableViewCell