Display Wind Outlook
Wind speed is displayed within a bar graph in each table view cell.
This helps with visualizing the wind forecast for the upcoming days.
Unfortunately the api call only returns the upcoming 48 hours so
everything after that is stubbed as of now.  The bar graph was created
using SwiftUI and is added to each UITableViewCell.  I also utilized
SwiftUI to create a PreviewProvider for my WindOutlookTableViewController
which has greatly reduced my development time when adjusting the UI.

Changes
- Create Bar Graph using SwiftUI to display wind speeds
- Display wind forecast in each TableViewCell